# Conflicts:
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/ksu/001/a/ksu001.a.vm.ts
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/program/ProgramIdConsts.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpLea15InfoCommandHandler.java
@@ -14,11 +14,11 @@ import nts.uk.shr.pereg.app.command.PeregUpdateCommandHandler;
 
 @Stateless
 public class UpdateSpLea15InfoCommandHandler extends CommandHandler<UpdateSpecialleave15informationCommand>
-implements PeregUpdateCommandHandler<UpdateSpecialleave15informationCommand>{
+		implements PeregUpdateCommandHandler<UpdateSpecialleave15informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler updateSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00053";
@@ -31,12 +31,13 @@ implements PeregUpdateCommandHandler<UpdateSpecialleave15informationCommand>{
 
 	@Override
 	protected void handle(CommandHandlerContext<UpdateSpecialleave15informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid,command.getSID(),SpecialLeaveCode.CS00053.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
-				command.getGrantDays()!=null? command.getGrantDays().intValue() : null, command.getGrantTable());
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00053.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
+				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		updateSpLeaInfoCommandHandler.updateHandler(domain);
 	}
