+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpLea1InfoCommandHandler.java
@@ -14,11 +14,11 @@ import nts.uk.shr.pereg.app.command.PeregUpdateCommandHandler;
 
 @Stateless
 public class UpdateSpLea1InfoCommandHandler extends CommandHandler<UpdateSpecialleave1informationCommand>
-implements PeregUpdateCommandHandler<UpdateSpecialleave1informationCommand>{
+		implements PeregUpdateCommandHandler<UpdateSpecialleave1informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler updateSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00025";
@@ -31,12 +31,13 @@ implements PeregUpdateCommandHandler<UpdateSpecialleave1informationCommand>{
 
 	@Override
 	protected void handle(CommandHandlerContext<UpdateSpecialleave1informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
+
 		String cid = AppContexts.user().companyId();
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid,command.getSID(),SpecialLeaveCode.CS00025.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
-				command.getGrantDays()!=null? command.getGrantDays().intValue() : null, command.getGrantTable());
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00025.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
+				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		updateSpLeaInfoCommandHandler.updateHandler(domain);
 	}
